requirements.txt
SQLAlchemy~=2.0.27
python-dotenv~=1.0.1
pandas~=2.2.1
py2neo~=2021.2.4

elasticsearch~=8.13.0
numpy~=1.26.4
pika~=1.3.2
django-scheduler~=0.10.1